requirements.txt
python-telegram-bot==21.6
flask==3.0.3
requests==2.32.3
gunicorn==22.0.0
